refactor FlowPlayer into folder components
