+++ src/pages/FlowPlayer/Skeleton.tsx
@@ -1,0 +1,53 @@
+import { Card, CardContent } from "../../components/ui/card";
+import { Skeleton } from "../../components/ui/skeleton";
+
+export default function PlayerSkeleton() {
+  return (
+    <div className="min-h-screen bg-gradient-to-br from-blue-50 via-white to-purple-50">
+      {/* Header Skeleton */}
+      <header className="bg-white/80 backdrop-blur-sm border-b">
+        <div className="max-w-4xl mx-auto px-4 py-4">
+          <div className="flex items-center justify-between">
+            <div className="flex items-center gap-4">
+              <Skeleton className="h-8 w-16" />
+              <div className="hidden sm:block space-y-2">
+                <Skeleton className="h-5 w-32" />
+                <Skeleton className="h-4 w-24" />
+              </div>
+            </div>
+            <div className="flex items-center gap-3">
+              <Skeleton className="h-6 w-12" />
+              <Skeleton className="h-4 w-8" />
+            </div>
+          </div>
+          <div className="mt-4">
+            <Skeleton className="h-2 w-full" />
+          </div>
+        </div>
+      </header>
+
+      {/* Main Content Skeleton */}
+      <main className="flex-1 flex items-center justify-center p-4 py-8">
+        <div className="w-full max-w-3xl">
+          <Card className="shadow-xl border-0">
+            <CardContent className="p-8 sm:p-12">
+              <div className="space-y-8">
+                <div className="text-center space-y-4">
+                  <Skeleton className="h-8 w-3/4 mx-auto" />
+                </div>
+                <div className="text-center">
+                  <Skeleton className="h-20 w-full" />
+                </div>
+                <div className="space-y-3">
+                  <Skeleton className="h-12 w-full" />
+                  <Skeleton className="h-12 w-full" />
+                  <Skeleton className="h-12 w-full" />
+                </div>
+              </div>
+            </CardContent>
+          </Card>
+        </div>
+      </main>
+    </div>
+  );
+}
